add classes Ui, Storage, Parser, TaskList
